aut est un long
long est un int 64-bit
ds 32-bit (int normal) ca fit pas 333555588888, mais ca fit ds 64-bit

+++ Projet1/app/src/main/java/ca/qc/cgodin/projet1/model/Succursale.kt
@@ -8,7 +8,7 @@ import java.io.Serializable
 @Entity(tableName = "succursale_table", primaryKeys = ["Aut", "Ville"])
 data class Succursale(
     @SerializedName("Aut")
-    val Aut: Int,
+    val Aut: Long,
     @SerializedName("Ville")
     val Ville: String,
     @SerializedName("Budget")
