+++ Projet1/app/src/main/java/ca/qc/cgodin/projet1/model/schema/AjoutSuccursaleSchema.kt
@@ -1,7 +1,7 @@
 package ca.qc.cgodin.projet1.model.schema
 
 data class AjoutSuccursaleSchema(
-    val Aut: Int,
+    val Aut: Long,
     val Ville: String,
     val Budget: Int
 )
