+++ Projet1/app/src/main/java/ca/qc/cgodin/projet1/model/schema/BudgetSuccursaleSchema.kt
@@ -1,6 +1,6 @@
 package ca.qc.cgodin.projet1.model.schema
 
 data class BudgetSuccursaleSchema(
-    val Aut: Int,
+    val Aut: Long,
     val Ville: String
 )
