+++ Projet1/app/src/main/java/ca/qc/cgodin/projet1/model/schema/CompteSuccursaleSchema.kt
@@ -1,5 +1,5 @@
 package ca.qc.cgodin.projet1.model.schema
 
 data class CompteSuccursaleSchema(
-    val Aut: Int
+    val Aut: Long
 )
